+++ src/DocumentContext.js
@@ -23,7 +23,7 @@ class DocumentContext extends EventEmitter {
 		this.addPage(pageSize);
 	}
 
-	beginColumnGroup() {
+	beginColumnGroup(marginXTopParent) {
 		this.snapshots.push({
 			x: this.x,
 			y: this.y,
@@ -41,6 +41,13 @@ class DocumentContext extends EventEmitter {
 		});
 
 		this.lastColumnWidth = 0;
+		if (marginXTopParent) {
+			this.marginXTopParent = marginXTopParent;
+		}
+	}
+
+	resetMarginXTopParent() {
+		this.marginXTopParent = null;
 	}
 
 	beginColumn(width, offset, endingCell) {
@@ -65,10 +72,10 @@ class DocumentContext extends EventEmitter {
 		}
 	}
 
-	markEnding(endingCell) {
+	markEnding(endingCell, originalXOffset, discountY) {
 		this.page = endingCell._columnEndingContext.page;
-		this.x = endingCell._columnEndingContext.x;
-		this.y = endingCell._columnEndingContext.y;
+		this.x = endingCell._columnEndingContext.x + originalXOffset;
+		this.y = endingCell._columnEndingContext.y - discountY;
 		this.availableWidth = endingCell._columnEndingContext.availableWidth;
 		this.availableHeight = endingCell._columnEndingContext.availableHeight;
 		this.lastColumnWidth = endingCell._columnEndingContext.lastColumnWidth;
@@ -128,14 +135,19 @@ class DocumentContext extends EventEmitter {
 	initializePage() {
 		this.y = this.pageMargins.top;
 		this.availableHeight = this.getCurrentPage().pageSize.height - this.pageMargins.top - this.pageMargins.bottom;
-		this.pageSnapshot().availableWidth = this.getCurrentPage().pageSize.width - this.pageMargins.left - this.pageMargins.right;
+		const { pageCtx, isSnapshot } = this.pageSnapshot();
+		pageCtx.availableWidth = this.getCurrentPage().pageSize.width - this.pageMargins.left - this.pageMargins.right;
+		if (isSnapshot && this.marginXTopParent) {
+			pageCtx.availableWidth -= this.marginXTopParent[0];
+			pageCtx.availableWidth -= this.marginXTopParent[1];
+		}
 	}
 
 	pageSnapshot() {
 		if (this.snapshots[0]) {
-			return this.snapshots[0];
+			return { pageCtx: this.snapshots[0], isSnapshot: true };
 		} else {
-			return this;
+			return { pageCtx: this, isSnapshot: false };
 		}
 	}
 
@@ -185,6 +197,16 @@ class DocumentContext extends EventEmitter {
 		let nextPageIndex = this.page + 1;
 		let prevPage = this.page;
 		let prevY = this.y;
+
+		// If we are in a column group
+		if (this.snapshots.length > 0) {
+			let lastSnapshot = this.snapshots[this.snapshots.length - 1];
+			// We have to update prevY accordingly by also taking into consideration
+			// the 'y' of cells that don't break page
+			if (lastSnapshot.bottomMost && lastSnapshot.bottomMost.y) {
+				prevY = Math.max(this.y, lastSnapshot.bottomMost.y);
+			}
+		}
 
 		let createNewPage = nextPageIndex >= this.pages.length;
 		if (createNewPage) {