+++ src/TableProcessor.js
@@ -108,7 +108,7 @@ class TableProcessor {
 		const headerRows = tableNode.table.headerRows;
 
 		if (isPositiveInteger(headerRows)) {
-      this.headerRows = headerRows;
+			this.headerRows = headerRows;
 
 			if (this.headerRows > tableNode.table.body.length) {
 				throw new Error(`Too few rows in the table. Property headerRows requires at least ${this.headerRows}, contains only ${tableNode.table.body.length}`);
@@ -119,20 +119,24 @@ class TableProcessor {
 			const keepWithHeaderRows = tableNode.table.keepWithHeaderRows;
 
 			if (isPositiveInteger(keepWithHeaderRows)) {
-					this.rowsWithoutPageBreak += keepWithHeaderRows;
+				this.rowsWithoutPageBreak += keepWithHeaderRows;
 			}
 		}
 
 		this.dontBreakRows = tableNode.table.dontBreakRows || false;
 
-		if (this.rowsWithoutPageBreak) {
+		if (this.rowsWithoutPageBreak || this.dontBreakRows) {
 			writer.beginUnbreakableBlock();
+			// Draw the top border of the table
+			this.drawHorizontalLine(0, writer);
+			if (this.rowsWithoutPageBreak && this.dontBreakRows) {
+				// We just increase the value of transactionLevel
+				writer.beginUnbreakableBlock();
+			}
 		}
 
 		// update the border properties of all cells before drawing any lines
 		prepareCellBorders(this.tableNode.table.body);
-
-		this.drawHorizontalLine(0, writer);
 	}
 
 	onRowBreak(rowIndex, writer) {
@@ -151,7 +155,12 @@ class TableProcessor {
 
 		this.rowCallback = this.onRowBreak(rowIndex, writer);
 		writer.addListener('pageChanged', this.rowCallback);
-		if (this.dontBreakRows) {
+		if (rowIndex == 0 && !this.dontBreakRows && !this.rowsWithoutPageBreak) {
+			// We store the 'y' to draw later and if necessary the top border of the table
+			this._tableTopBorderY = writer.context().y;
+			writer.context().moveDown(this.topLineWidth);
+		}
+		if (this.dontBreakRows && rowIndex > 0) {
 			writer.beginUnbreakableBlock();
 		}
 		this.rowTopY = writer.context().y;
@@ -162,7 +171,7 @@ class TableProcessor {
 		writer.context().moveDown(this.rowPaddingTop);
 	}
 
-	drawHorizontalLine(lineIndex, writer, overrideY) {
+	drawHorizontalLine(lineIndex, writer, overrideY, moveDown = true, forcePage) {
 		let lineWidth = this.layout.hLineWidth(lineIndex, this.tableNode);
 		if (lineWidth) {
 			let style = this.layout.hLineStyle(lineIndex, this.tableNode);
@@ -260,7 +269,7 @@ class TableProcessor {
 							lineWidth: lineWidth,
 							dash: dash,
 							lineColor: borderColor
-						}, false, overrideY);
+						}, false, isNumber(overrideY), null, forcePage);
 						currentLine = null;
 						borderColor = null;
 						cellAbove = null;
@@ -270,7 +279,9 @@ class TableProcessor {
 				}
 			}
 
-			writer.context().moveDown(lineWidth);
+			if (moveDown) {
+				writer.context().moveDown(lineWidth);
+			}
 		}
 	}
 
@@ -406,6 +417,15 @@ class TableProcessor {
 		ys[ys.length - 1].y1 = endingY;
 
 		let skipOrphanePadding = (ys[0].y1 - ys[0].y0 === this.rowPaddingTop);
+		if (rowIndex === 0 && !skipOrphanePadding && !this.rowsWithoutPageBreak && !this.dontBreakRows) {
+			// Draw the top border of the table
+			let pageTableStartedAt = null;
+			if (pageBreaks && pageBreaks.length > 0) {
+				// Get the page where table started at
+				pageTableStartedAt = pageBreaks[0].prevPage;
+			}
+			this.drawHorizontalLine(0, writer, this._tableTopBorderY, false, pageTableStartedAt);
+		}
 		for (let yi = (skipOrphanePadding ? 1 : 0), yl = ys.length; yi < yl; yi++) {
 			let willBreak = yi < ys.length - 1;
 			let rowBreakWithoutHeader = (yi > 0 && !this.headerRows);
@@ -423,6 +443,14 @@ class TableProcessor {
 				//TODO: buggy, availableHeight should be updated on every pageChanged event
 				// TableProcessor should be pageChanged listener, instead of processRow
 				this.reservedAtBottom = 0;
+			}
+
+			// Draw horizontal lines before the vertical lines so they are not overridden
+			if (willBreak && this.layout.hLineWhenBroken !== false) {
+				this.drawHorizontalLine(rowIndex + 1, writer, y2);
+			}
+			if (rowBreakWithoutHeader && this.layout.hLineWhenBroken !== false) {
+				this.drawHorizontalLine(rowIndex, writer, y1);
 			}
 
 			for (let i = 0, l = xs.length; i < l; i++) {
@@ -490,7 +518,9 @@ class TableProcessor {
 								h: bgHeight,
 								lineWidth: 0,
 								color: fillColor,
-								fillOpacity: fillOpacity
+								fillOpacity: fillOpacity,
+								// mark if we are in an unbreakable block
+								_isFillColorFromUnbreakable: !!writer.transactionLevel
 							}, false, true, writer.context().backgroundLength[writer.context().page]);
 						}
 
@@ -508,13 +538,6 @@ class TableProcessor {
 						}
 					}
 				}
-			}
-
-			if (willBreak && this.layout.hLineWhenBroken !== false) {
-				this.drawHorizontalLine(rowIndex + 1, writer, y2);
-			}
-			if (rowBreakWithoutHeader && this.layout.hLineWhenBroken !== false) {
-				this.drawHorizontalLine(rowIndex, writer, y1);
 			}
 		}
 
@@ -554,7 +577,8 @@ class TableProcessor {
 
 		if (this.dontBreakRows) {
 			const pageChangedCallback = () => {
-				if (!this.headerRows && this.layout.hLineWhenBroken !== false) {
+				if (rowIndex > 0 && !this.headerRows && this.layout.hLineWhenBroken !== false) {
+					// Draw the top border of the row after a page break
 					this.drawHorizontalLine(rowIndex, writer);
 				}
 			};